part4 completed

+++ part4/blog_list/controllers/blogs.js
@@ -1,23 +1,20 @@
 const blogsRouter = require('express').Router();
+const jwt = require('jsonwebtoken')
+require('dotenv').config()
 const Blog = require('../models/blog');
-const logger = require('../utils/logger');
+const User = require('../models/user')
+const logger = require('../utils/logger')
 
 
 blogsRouter.get('/', async (request, response) => {
-    const blogs = await Blog.find({})
+    const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
     response.json(blogs)
 });
 
-blogsRouter.get('/info', async (request, response) => {
-  const blogs = await Blog.find({})
-  response.json({cantidad: blogs.length})
-});
-
 blogsRouter.get('/:id', async (request, response) => {
-  const blogs = await  Blogs.findById(request.params.id);
-
-  if (blogs) {
-    response.json(blogs);
+  const blog = await Blog.findById(request.params.id);
+  if (blog) {
+    response.json(blog);
   } else {
     response.status(404).end();
   }
@@ -27,14 +24,27 @@ blogsRouter.post('/', async (request, response) => {
   
  const { title1, author1, url1, likes1 } = request.body;
 
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+  if (!request.token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+
+  const user = await User.findById(decodedToken.id)
+
   blog = new Blog({
     title: title1,
     author: author1,
     url: url1,
-    likes:likes1 
+    likes:likes1,
+    user: user._id
   })
   
   const savedBlog = await blog.save()
+
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
+
   response.json(savedBlog)
   
 });
@@ -48,8 +58,21 @@ blogsRouter.put('/:id', async (request, response) => {
 });
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end();
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+  if (!request.token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+
+  const user = await User.findById(decodedToken.id)
+  const blog = await Blog.findById(request.params.id)
+
+  if(blog.user.toString() === user._id.toString()){
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end();
+  }
+
 });
 
 module.exports = blogsRouter
